refactor: 모든 Page 폴더안에 넣도록 수정
